extensions/StockListUpdateChecker/src/requirements.txt
onepush==1.1.1
pandas==1.4.0
python-dotenv==0.19.2
tushare==1.2.83